Challenge texts and attacks
So this change is minor but important. putting challenge texts from enemies.
There is a bug if we try to display text permanently on top of an enemy which you can readily see if you fight a turnip.
We also have some attacks displayed that can be set individually (necessary if you want attack variety later), which could be combined with a trigger system for a nicer menu but for the moment the default menu is still used.